Pipeline is disabled upon TX error
