+++ zenoh/src/net/transport/multicast/link.rs
@@ -113,7 +113,7 @@ impl TransportLinkMulticast {
             let c_transport = self.transport.clone();
             let handle = task::spawn(async move {
                 let res = tx_task(
-                    pipeline,
+                    pipeline.clone(),
                     c_link.clone(),
                     config,
                     initial_sns,
@@ -121,6 +121,7 @@ impl TransportLinkMulticast {
                     c_transport.stats.clone(),
                 )
                 .await;
+                pipeline.disable();
                 if let Err(e) = res {
                     log::debug!("{}", e);
                     // Spawn a task to avoid a deadlock waiting for this same task
