+++ zenoh/src/net/transport/unicast/link.rs
@@ -90,13 +90,14 @@ impl TransportLinkUnicast {
             let c_transport = self.transport.clone();
             let handle = task::spawn(async move {
                 let res = tx_task(
-                    pipeline,
+                    pipeline.clone(),
                     c_link.clone(),
                     keep_alive,
                     #[cfg(feature = "stats")]
                     c_transport.stats.clone(),
                 )
                 .await;
+                pipeline.disable();
                 if let Err(e) = res {
                     log::debug!("{}", e);
                     // Spawn a task to avoid a deadlock waiting for this same task
